requirements.txt
earthengine-api
geemap
rasterio
pillow
numpy
label-studio
boto3
python-dotenv

